refactor: move aggregators to their own module
